Problem: msgspec conversion belongs in messaging
Solution: Move it out from the prototype output reduction
